+++ CHANGES.rst
@@ -1,3 +1,15 @@
+v1.0
+----
+- added more output information to fits file
+   - lightcurve properties
+   - rise time
+- change in GUI design
+   - added frame for list of stars
+   - added ability to select bad data
+   - added indicators for rise time
+- split out certain functions from GUI class
+- general cleaning of code
+
 v0.3
 ----
 - compatibility with lightkurve v2.0
@@ -6,7 +18,7 @@ v0.2
 ----
 - added a Blazhko check option
 - added an aperture size selection
-- changed GUI
+- changed GUI code
    - added frames
    - adjusted options